nicer info on parsing, fix encode bug

+++ download_planet_notes.sh
@@ -12,6 +12,12 @@ $WGET "$URL_BASE/$PLANET_NOTES_MD5"  || exit 2
 md5sum --status -c $PLANET_NOTES_MD5 || exit 3
 
 echo "$URL_BASE/$PLANET_NOTES_FILE: `ls -l $PLANET_NOTES_FILE`"
-cp -a "$PLANET_NOTES_FILE" "OK.$PLANET_NOTES_FILE"
-exec time ./myosmnotes_parser.pl
-exit 99
+if [ ! -e "OK.$PLANET_NOTES_FILE" -o "$PLANET_NOTES_FILE" -nt "OK.$PLANET_NOTES_FILE" ]
+then
+	cp -a "$PLANET_NOTES_FILE" "OK.$PLANET_NOTES_FILE"
+	exec time ./myosmnotes_parser.pl
+	exit 99
+else
+	echo "No updates."
+	exit 0
+fi
